+++ backend/src/main/java/com/ody/mate/domain/controller/MateControllerSwagger.java
@@ -1,7 +1,7 @@
 package com.ody.mate.domain.controller;
 
 import com.ody.mate.domain.dto.MateSaveRequest;
-import com.ody.meeting.dto.MeetingSaveResponse;
+import com.ody.meeting.dto.response.MeetingSaveResponse;
 import com.ody.swagger.annotation.ErrorCode400;
 import com.ody.swagger.annotation.ErrorCode401;
 import com.ody.swagger.annotation.ErrorCode500;
