+++ backend/src/main/java/com/ody/meeting/dto/response/MateResponse.java
@@ -1,4 +1,4 @@
-package com.ody.meeting.dto;
+package com.ody.meeting.dto.response;
 
 public record MateResponse(String nickname) {
